fix(gate-3b): extract graph_read test module — 600→306 lines (#608)
